Add a toplevel option to generate grpc-rs bindings (see notes)
This commit
https://github.com/tikv/grpc-rs/commit/8da61263fee86926ad45bc68c81cbe94acb1eec2
changed the grpc-sys build such that cross compile builds for 32-bit arm
cannot use the bindings.rs packaged with the crate, but that commit also
makes it such that the new bindings are not generated as part of the
build. This results in a build failure in buildroot similar to this:

    error: couldn't read
    $directory/target/armv7-unknown-linux-gnueabihf/release/build/
        grpcio-sys-2aca5deabe523c8e/out/grpc-bindings.rs:
    No such file or directory (os error 2)
     -->
     /home/user/.cargo/registry/src/index.crates.io-6f17d22bba15001f/
        grpcio-sys-0.12.1+1.46.5-patched/src/lib.rs:8:5
       |
       8 |     include!(env!("BINDING_PATH"));

adding this feature to the top level allows a consumer to enable the
bindings generation as part of the build.

Becuase of the way the code is set up in grpc-sys/build.rs, we can
enable this feature on all platforms and the bindings will only be
generated when the `target_arch=arm` - not for x86 builds or for
aarch64 builds.